Upgrage Shop + Etape
- Le addSlot du shop peut maintenant prendre un float en multiplicateur. Cela permet de grimper trop vite en prix. On pourrait faire une main avec un nombre de slot égale à celui des players. On devra alors y mettre une scroll bar.
- Le shop possède des fonctions AddGold(int) et SubstractGold(int) qui vont ajouter ou retirer des golds et mettre a jour le widget de la main
- Les etapes doivent maintenant hérité de AStateGame. Le GameManager prends une liste de AStateGame et va faire selon l'ordre des étapes dans la liste. Les timers des étapes sont dans les StatesManager. Ils ont aussi un boolean si ils ont besoin du canva (War n'en aura pas besoin)

+++ Assets/Script/TheoScript/CardShop.cs
@@ -12,7 +12,7 @@ public class CardShop : AContainsSlots, IDropHandler, IPointerEnterHandler, IPoi
     public TextMeshProUGUI textForWidgetGold;
     public int nbSlotShop;
     public int addSlotShopPrice;
-    public int multiplicatorSlotShopPrice;
+    public float multiplicatorSlotShopPrice;
     public int gold;
     public int gainGoldIfWin;
     public int gainGoldIfLose;
